+++ LeanAndMeanCards/Cards/CardRegistration.cs
@@ -1,3 +1,4 @@
+using System;
 using LeanAndMeanCards.Utils;
 using UnboundLib.Cards;
 
@@ -7,22 +8,33 @@ namespace LeanAndMeanCards.Cards
     {
         internal static void RegisterAll()
         {
-            CustomCard.BuildCard<Thief>(info => Thief.Card = info);
-            CustomCard.BuildCard<Takebacksies>(info => Takebacksies.Card = info);
+            Bind<Thief>(info => Thief.Card = info);
+            Bind<Takebacksies>(info => Takebacksies.Card = info);
             TakebacksiesInjector.Register();
-            CustomCard.BuildCard<SandbagSimulator>(info => SandbagSimulator.Card = info);
-            CustomCard.BuildCard<JarOfDirt>(info => JarOfDirt.Card = info);
-            CustomCard.BuildCard<Confetti>(info => Confetti.Card = info);
-            CustomCard.BuildCard<Shove>(info => Shove.Card = info);
-            CustomCard.BuildCard<Pisser>(info => Pisser.Card = info);
-            CustomCard.BuildCard<Doorstop>(info => Doorstop.Card = info);
-            CustomCard.BuildCard<BozoShoes>(info => BozoShoes.Card = info);
-            CustomCard.BuildCard<DraftSniper>(info => DraftSniper.Card = info);
-            CustomCard.BuildCard<SilverEgg>(info => SilverEgg.Card = info);
-            CustomCard.BuildCard<YeetCannon>(info => YeetCannon.Card = info);
-            CustomCard.BuildCard<Dynamite>(info => Dynamite.Card = info);
-            CustomCard.BuildCard<TaserTaserTaser>(info => TaserTaserTaser.Card = info);
-            CustomCard.BuildCard<SafetyNet>(info => SafetyNet.Card = info);
+            Bind<SandbagSimulator>(info => SandbagSimulator.Card = info);
+            Bind<JarOfDirt>(info => JarOfDirt.Card = info);
+            Bind<Confetti>(info => Confetti.Card = info);
+            Bind<Shove>(info => Shove.Card = info);
+            Bind<Pisser>(info => Pisser.Card = info);
+            Bind<Doorstop>(info => Doorstop.Card = info);
+            Bind<BozoShoes>(info => BozoShoes.Card = info);
+            Bind<DraftSniper>(info => DraftSniper.Card = info);
+            Bind<SilverEgg>(info => SilverEgg.Card = info);
+            Bind<YeetCannon>(info => YeetCannon.Card = info);
+            Bind<Dynamite>(info => Dynamite.Card = info);
+            Bind<TaserTaserTaser>(info => TaserTaserTaser.Card = info);
+            Bind<SafetyNet>(info => SafetyNet.Card = info);
+        }
+
+        private static void Bind<T>(Action<CardInfo> setStatic) where T : CustomCard
+        {
+            CustomCard.BuildCard<T>(info =>
+            {
+                setStatic(info);
+                // Unbound sets cardArt and cardName after SetupCard. The callback
+                // is the first moment both exist on the prefab.
+                CardArtFactory.TryAssignSprite(info);
+            });
         }
     }
 }